back-end/web-api/requirements.txt
python-dotenv
psycopg2-binary
flask
flask-sqlalchemy
flask-cors